requirements.txt
tensorflow
numpy
pandas
matplotlib
datetime
websocket
talib
pandas_datareader 
python-binance